feat: add assertion env vars
